# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/calendars:
    get:
      tags:
        - Calendars
      summary: List Calendars
      description: Get a paginated list of Calendars with optional filtering and sorting
      operationId: listCalendars
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Calendars
      summary: Create Calendar
      description: Create a new Calendar
      operationId: createCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/calendar_levels:
    get:
      tags:
        - CalendarLevels
      summary: List CalendarLevels
      description: Get a paginated list of CalendarLevels with optional filtering and sorting
      operationId: listCalendarLevels
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarLevelList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarLevels
      summary: Create CalendarLevel
      description: Create a new CalendarLevel
      operationId: createCalendarLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarLevelRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarLevel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Calendar:
      type: object
      required:
        - id
        - company_id
        - name
        - date_start
        - date_end
        - is_holiday
        - can_everyone_view
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
        date_start:
          type: string
          format: date
        date_end:
          type: string
          format: date
        is_holiday:
          type: boolean
        can_everyone_view:
          type: boolean
        note:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateCalendarRequest:
      type: object
      required:
        - company_id
        - name
        - date_start
        - date_end
        - is_holiday
        - can_everyone_view
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
        date_start:
          type: string
          format: date
        date_end:
          type: string
          format: date
        is_holiday:
          type: boolean
        can_everyone_view:
          type: boolean
        note:
          nullable: true
          type: string
        metadata:
          type: object
...
    CalendarList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Calendar'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    CalendarLevel:
      type: object
      required:
        - id
        - calendar_id
        - level_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        calendar_id:
          type: string
          format: uuid
        level_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCalendarLevelRequest:
      type: object
      required:
        - calendar_id
        - level_id
        - metadata
      properties:
        calendar_id:
          type: string
          format: uuid
        level_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CalendarLevelList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarLevel'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []